weather displaying done

+++ app/src/main/java/com/dbohdanov/weatherapp/repository/data_models/WeatherResponseModel.java
@@ -186,7 +186,7 @@ public class WeatherResponseModel {
         private double sea_level;
         private double grnd_level;
         private int humidity;
-        private int temp_kf;
+        private double temp_kf;
 
         public double getTemp() {
             return temp;
@@ -244,11 +244,11 @@ public class WeatherResponseModel {
             this.humidity = humidity;
         }
 
-        public int getTemp_kf() {
+        public double getTemp_kf() {
             return temp_kf;
         }
 
-        public void setTemp_kf(int temp_kf) {
+        public void setTemp_kf(double temp_kf) {
             this.temp_kf = temp_kf;
         }
     }
